automation_tools/requirements.txt
tqdm==4.60.0
requests==2.22.0
requests_toolbelt==0.9.1
python_magic==0.4.22
